+++ src/card/sets/y2001/mod.rs
@@ -3,5 +3,6 @@
 pub(super) mod apocalypse;
 pub(super) mod odyssey;
 pub(super) mod planeshift;
+pub(super) mod seventh_edition;
 
 pub(super) use super::{CardRecord, PrintingRecord};